# engine/docs2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flapjack API",
    "description": "Drop-in replacement for Algolia search. Provides full-text search with filters, facets, geo-search, and more.",
    "contact": {
      "name": "Flapjack",
      "url": "https://github.com/stuartcrobinson/flapjack"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7700",
      "description": "Local development"
    },
    {
      "url": "https://fj-us-west-1.flapjack.foo",
      "description": "Production US West"
    }
  ],
  "paths": {
...
    "/1/configs": {
      "get": {
        "tags": [
          "query-suggestions"
        ],
        "summary": "GET /1/configs — list all Query Suggestions configurations",
        "operationId": "list_configs",
        "responses": {
          "200": {
            "description": "All query suggestions configurations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QsConfig"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Query suggestions store read failed"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "query-suggestions"
        ],
        "summary": "Create a new query suggestions configuration and schedule the initial index build.",
        "operationId": "create_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QsConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Configuration created and build scheduled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QsMutationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query suggestions configuration"
          },
          "409": {
            "description": "Configuration already exists"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/1/configs/{indexName}": {
      "get": {
        "tags": [
          "query-suggestions"
        ],
        "summary": "GET /1/configs/:indexName — get a single configuration",
        "operationId": "get_config",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Query suggestions configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QsConfig"
                }
              }
            }
          },
          "400": {
            "description": "Invalid index name"
          },
          "404": {
            "description": "Configuration not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "query-suggestions"
        ],
        "summary": "Update an existing query suggestions configuration.",
        "operationId": "update_config",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QsConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Configuration updated and build scheduled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QsMutationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid index name or configuration"
          },
          "404": {
            "description": "Configuration not found"
          },
          "409": {
            "description": "Build already in progress"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "query-suggestions"
        ],
        "summary": "DELETE /1/configs/:indexName — delete configuration (does NOT delete the suggestions index)",
        "operationId": "delete_config",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QsMutationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid index name"
          },
          "404": {
            "description": "Configuration not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/1/configs/{indexName}/build": {
      "post": {
        "tags": [
          "query-suggestions"
        ],
        "summary": "Manually trigger a query suggestions index build for the specified configuration.",
        "operationId": "trigger_build",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Build triggered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QsMutationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid index name"
          },
          "404": {
            "description": "Configuration not found"
          },
          "409": {
            "description": "Build already in progress"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/1/configs/{indexName}/status": {
      "get": {
        "tags": [
          "query-suggestions"
        ],
        "summary": "GET /1/configs/:indexName/status — build status",
        "operationId": "get_status",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current build status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildStatus"
                }
              }
            }
          },
          "400": {
            "description": "Invalid index name"
          },
          "404": {
            "description": "Configuration not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/1/logs/{indexName}": {
      "get": {
        "tags": [
          "query-suggestions"
        ],
        "summary": "GET /1/logs/:indexName — build logs",
        "operationId": "get_logs",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Build logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogEntry"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid index name"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuildStatus": {
        "type": "object",
        "required": [
          "indexName"
        ],
        "properties": {
          "indexName": {
            "type": "string"
          },
          "isRunning": {
            "type": "boolean"
          },
          "lastBuiltAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastSuccessfulBuiltAt": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "LogEntry": {
        "type": "object",
        "required": [
          "timestamp",
          "level",
          "message",
          "contextLevel"
        ],
        "properties": {
          "contextLevel": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "level": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "QsConfig": {
        "type": "object",
        "required": [
          "indexName",
          "sourceIndices"
        ],
        "properties": {
          "allowSpecialCharacters": {
            "type": "boolean"
          },
          "enablePersonalization": {
            "type": "boolean"
          },
          "exclude": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "indexName": {
            "type": "string"
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourceIndices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QsSourceIndex"
            }
          }
        }
      },
      "QsFacet": {
        "type": "object",
        "required": [
          "attribute",
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "attribute": {
            "type": "string"
          }
        }
      },
      "QsMutationResponse": {
        "type": "object",
        "description": "Standard success response for query-suggestions mutation operations.",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "QsSourceIndex": {
        "type": "object",
        "required": [
          "indexName"
        ],
        "properties": {
          "analyticsTags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "facets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QsFacet"
            }
          },
          "generate": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "indexName": {
            "type": "string"
          },
          "minHits": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "minLetters": {
            "type": "integer",
            "minimum": 0
          },
          "replicas": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-algolia-api-key"
      }
    }
  },
...
}